Used source_path in config file.
